[TECH](RC)(WarningNotice) rename src-fixtures file to match folder and easier navigation
